rename logging.py to loggers.py to avoid import confusion
